+++ src/uu/chgrp/src/chgrp.rs
@@ -11,22 +11,17 @@
 extern crate uucore;
 pub use uucore::entries;
 use uucore::fs::resolve_relative_path;
-use uucore::libc::{self, gid_t, lchown};
+use uucore::libc::gid_t;
+use uucore::perms::{wrap_chgrp, Verbosity};
 
 extern crate walkdir;
 use walkdir::WalkDir;
-
-use std::io::Error as IOError;
-use std::io::Result as IOResult;
 
 use std::fs;
 use std::fs::Metadata;
 use std::os::unix::fs::MetadataExt;
 
 use std::path::Path;
-
-use std::ffi::CString;
-use std::os::unix::ffi::OsStrExt;
 
 static SYNTAX: &str =
     "chgrp [OPTION]... GROUP FILE...\n or :  chgrp [OPTION]... --reference=RFILE FILE...";
@@ -165,14 +160,6 @@ pub fn uumain(args: impl uucore::Args) -> i32 {
     executor.exec()
 }
 
-#[derive(PartialEq, Debug)]
-enum Verbosity {
-    Silent,
-    Changes,
-    Verbose,
-    Normal,
-}
-
 struct Chgrper {
     dest_gid: gid_t,
     bit_flag: u8,
@@ -199,23 +186,6 @@ impl Chgrper {
             ret |= self.traverse(f);
         }
         ret
-    }
-
-    fn chgrp<P: AsRef<Path>>(&self, path: P, dgid: gid_t, follow: bool) -> IOResult<()> {
-        let path = path.as_ref();
-        let s = CString::new(path.as_os_str().as_bytes()).unwrap();
-        let ret = unsafe {
-            if follow {
-                libc::chown(s.as_ptr(), (0 as gid_t).wrapping_sub(1), dgid)
-            } else {
-                lchown(s.as_ptr(), (0 as gid_t).wrapping_sub(1), dgid)
-            }
-        };
-        if ret == 0 {
-            Ok(())
-        } else {
-            Err(IOError::last_os_error())
-        }
     }
 
     #[cfg(windows)]
@@ -269,7 +239,13 @@ impl Chgrper {
             }
         }
 
-        let ret = self.wrap_chgrp(path, &meta, follow_arg);
+        let ret = wrap_chgrp(
+            path,
+            &meta,
+            self.dest_gid,
+            follow_arg,
+            self.verbosity.clone(),
+        );
 
         if !self.recursive {
             ret
@@ -297,7 +273,7 @@ impl Chgrper {
                 }
             };
 
-            ret = self.wrap_chgrp(path, &meta, follow);
+            ret = wrap_chgrp(path, &meta, self.dest_gid, follow, self.verbosity.clone());
         }
         ret
     }
@@ -323,51 +299,5 @@ impl Chgrper {
             })
         };
         Some(meta)
-    }
-
-    fn wrap_chgrp<P: AsRef<Path>>(&self, path: P, meta: &Metadata, follow: bool) -> i32 {
-        use self::Verbosity::*;
-        let mut ret = 0;
-        let dest_gid = self.dest_gid;
-        let path = path.as_ref();
-        if let Err(e) = self.chgrp(path, dest_gid, follow) {
-            match self.verbosity {
-                Silent => (),
-                _ => {
-                    show_info!("changing group of '{}': {}", path.display(), e);
-                    if self.verbosity == Verbose {
-                        println!(
-                            "failed to change group of {} from {} to {}",
-                            path.display(),
-                            entries::gid2grp(meta.gid()).unwrap(),
-                            entries::gid2grp(dest_gid).unwrap()
-                        );
-                    };
-                }
-            }
-            ret = 1;
-        } else {
-            let changed = dest_gid != meta.gid();
-            if changed {
-                match self.verbosity {
-                    Changes | Verbose => {
-                        println!(
-                            "changed group of {} from {} to {}",
-                            path.display(),
-                            entries::gid2grp(meta.gid()).unwrap(),
-                            entries::gid2grp(dest_gid).unwrap()
-                        );
-                    }
-                    _ => (),
-                };
-            } else if self.verbosity == Verbose {
-                println!(
-                    "group of {} retained as {}",
-                    path.display(),
-                    entries::gid2grp(dest_gid).unwrap()
-                );
-            }
-        }
-        ret
     }
 }
