+++ src/uucore/Cargo.toml
@@ -43,6 +43,7 @@ entries = ["libc"]
 fs = ["libc"]
 mode = ["libc"]
 parse_time = []
+perms = ["libc"]
 process = ["libc"]
 signals = []
 utf8 = []
